pyproject.toml
[project]
name = "genderlm"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.121.1",
    "gradio>=3.36.1",
    "insightface>=0.7.3",
    "onnxruntime>=1.23.2",
    "pillow>=12.0.0",
    "python-multipart>=0.0.20",
    "torch>=2.9.1",
    "transformers>=4.57.1",
    "uvicorn>=0.38.0",
]
